Add pipeline with an adaptive cascade, and correct an unsupported claim
Testing the cascade against real Sentinel-2 data contradicted the design doc, so
both the implementation and the documentation changed.

What real data showed. On a 1920x1920 chip of pure open ocean off Accra
(S2A_MSIL2A_20240527T100601_R022_T30NZM, 100% water, no known debris), FDI over
water measured mean +0.00178, p50 +0.00157, p99 +0.00917, p99.9 +0.01316. The
fixed threshold of 0.006 therefore flagged 6.35% of pure water as candidate and
accepted every tile. The cascade would have appeared to work while saving nothing.

The fix. The cutoff is now taken from a high percentile of each scene's own water
FDI distribution, since FDI magnitude moves with atmospheric correction, sun angle
and water type, and no constant transfers between scenes. The old constant is kept
as an absolute floor. On the same chip this cut candidate pixels from 234,262 to
3,686. Acceptance also now requires a connected region of at least 8 px, because
scattered single pixels are sun glint while real debris is spatially coherent.

Corrected claim. The README and design doc said the cascade cuts 40 minutes "to
minutes". Measured on a realistic coastal chip (53.7% water, 13.1% cloud, 36
tiles) it skips 16 of 36 tiles, 44% of detector calls, 11.1 min to 6.2 min. Both
documents now state the measured figure and note that the saving is largest on
coastal and cloudy scenes and smallest on open ocean.

Also documents the limitation inherent to percentile thresholding: debris filling
more than (100 - percentile)% of a tile's water would raise the cutoff into its
own values. At p99.9 that is roughly 9 hectares of continuous material.

19 cascade tests. One caught a physics error in its own fixture: planted debris
with NIR 0.075 against green 0.025 gives NDWI -0.5, so it was correctly rejected
as not being water. Real debris at 10 m is a sub-pixel mixture.

+++ src/mdebris/geo/raster.py
@@ -84,7 +84,9 @@ def _resampling(method: str | Resampling) -> Resampling:
         return Resampling[str(method).lower()]
     except KeyError as exc:
         options = ", ".join(sorted(r.name for r in Resampling))
-        raise ValueError(f"unknown resampling method {method!r}; expected one of {options}") from exc
+        raise ValueError(
+            f"unknown resampling method {method!r}; expected one of {options}"
+        ) from exc
 
 
 def window_transform(window: WindowLike, source: Affine | str | Path) -> Affine:
@@ -201,7 +203,9 @@ def read_bands(
         ref_win = _as_window(window) if window is not None else Window(0, 0, ref.width, ref.height)
         ref_crs = ref.crs
         geo_bounds = rio_windows.bounds(ref_win, ref.transform)
-    shape = target_shape or (int(round(ref_win.height)), int(round(ref_win.width)))
+    # Window dimensions can be numpy floats, so round through float to land on a
+    # plain int that rasterio's out_shape accepts.
+    shape = target_shape or (round(float(ref_win.height)), round(float(ref_win.width)))
 
     out: dict[str, np.ndarray] = {}
     for name, href in hrefs.items():
